@@ -1,6 +1,6 @@
 import React, { FunctionComponent } from 'react';
 import { NavLink as ReactRouterNavLink } from "react-router-dom";
-import { route } from 'react-ts-boiler';
+import { route } from 'buro-lib-ts';
 import NavItem from './NavItem';
 
 const NavBar: FunctionComponent = () => {
@@ -8,7 +8,7 @@ const NavBar: FunctionComponent = () => {
     return (
         <nav className={'navbar'}>
             <div className={'navbar__content'}>
-                <ReactRouterNavLink exact={true} to={route('home')}>
+                <ReactRouterNavLink to={route('home')}>
                     <div className={"navbar__container"}>
                         <h1>Toetsweb</h1>
                     </div>
@@ -16,15 +16,15 @@ const NavBar: FunctionComponent = () => {
                 
 
                 <div className='navbar__content--right'>
-                    <ReactRouterNavLink exact={true} to={route('home')}>
+                    <ReactRouterNavLink to={route('home')}>
                         <NavItem item={'Het toetsweb'} color={'purple'}/>
                     </ReactRouterNavLink>
 
-                    <ReactRouterNavLink exact={true} to={route('scan')}>
+                    <ReactRouterNavLink to={route('scan')}>
                         <NavItem item={'Toetsweb scan'} color={'orange'}/>
                     </ReactRouterNavLink>
 
-                    <ReactRouterNavLink exact={true} to={'/result'}>
+                    <ReactRouterNavLink to={'/result'}>
                         <NavItem item={'Resultaat'} color={'green'}/>
                     </ReactRouterNavLink>
                 </div>
