@@ -1,5 +1,5 @@
 import React, { FunctionComponent } from 'react';
-import {Page, route} from 'react-ts-boiler';
+import {Page, route} from 'buro-lib-ts';
 import { NavLink as ReactRouterNavLink } from "react-router-dom";
 
 interface Props {}
@@ -9,7 +9,7 @@ const NotFoundPage: FunctionComponent<Props> = () => {
     return (
         <Page className={'not-found-page'}>
             <h2>404 Pagina niet gevonden</h2>
-            <ReactRouterNavLink exact={true} to={route('home')}>
+            <ReactRouterNavLink to={route('home')}>
                 <div className='not-found-page__button'>
                     <p>Terug naar home</p>
                 </div>
