@@ -1,9 +1,12 @@
 import React, {FunctionComponent, useEffect} from 'react';
-import {Page} from "react-ts-boiler";
+import {Page} from 'buro-lib-ts';
 import Button from "../layout/Button";
 import ResultFragment from "../layout/ResultFragment";
 import scanData from "../../assets/data/scandata.json";
 import downloadFile from "../../utils/FileDownloader";
+import JSZip from 'jszip';
+
+const saveAs = require('save-svg-as-png');
 
 interface Props {
 }
@@ -13,10 +16,6 @@ const Result: FunctionComponent<Props> = () => {
     const entities = scanData.entities;
 
     useEffect(() => {
-
-        console.log(entities);
-        
-
         entities.forEach((entity, entityIndex) => {
             entity.elements.forEach((element, elementIndex) => {
                 const rawAnswer = window.localStorage.getItem(`${entityIndex}.${elementIndex}`);
@@ -27,13 +26,14 @@ const Result: FunctionComponent<Props> = () => {
                     window.location.href = '/scan';
                 }
             })
-        })
+        });
     }, []);
 
     const getPositionResult = (entity: number, element: number) => {
         const answer = JSON.parse(
             window.localStorage.getItem(`${entity}.${element}`) as string
-        )
+        );
+
         return answer.checkedPositie + 1;
     }
 
@@ -64,7 +64,8 @@ const Result: FunctionComponent<Props> = () => {
     }
 
     const downloadResults = () => {
-        let fileData = ""
+        const zip = new JSZip();
+        let fileData = "";
 
         entities.forEach((entity, entityIndex) => {
             fileData += entity.name + "\n";
@@ -78,17 +79,27 @@ const Result: FunctionComponent<Props> = () => {
             fileData += "\n";
         });
 
+        zip.file('Resultaten.txt', fileData);
 
+        const icons: Element[] = [];
+        document.querySelectorAll('.svg-model').forEach(svg => icons.push(svg));
 
-        downloadFile(new Blob([fileData]), 'resultaten.txt');
+        Promise.all(icons.map(svg => saveAs.svgAsPngUri(svg))).then(([one, two]) => {
+            zip.file('Positie.png', one.replace(/^data:image\/(png|jpg);base64,/, ""), { base64: true });
+            zip.file('Ambitie.png', two.replace(/^data:image\/(png|jpg);base64,/, ""), { base64: true });
+
+            zip.generateAsync({ type: 'blob' }).then(content => {
+                downloadFile(content, 'Resultaat.zip');
+            });
+        });
     }
 
     return (
         <Page className='result'>
             <h1 className='result__title'>Resultaat</h1>
             <div className='result__container'>
-                <ResultFragment pageTitle={'Positie'} getResult={getPositionResult} getFeedback={getPositionFeedback}/>
-                <ResultFragment pageTitle={'Ambitie'} getResult={getAmbitionResult} getFeedback={getAmbitionFeedback}/>
+                <ResultFragment pageTitle={'Positie'} getResult={getPositionResult} getFeedback={getPositionFeedback} />
+                <ResultFragment pageTitle={'Ambitie'} getResult={getAmbitionResult} getFeedback={getAmbitionFeedback} />
             </div>
             <div className='result__download-container'>
                 <div className='result__download-button'>
